requirements.txt
# ================================
# 核心框架
# ================================
langchain==1.0.0
langchain-community==0.4.1
langchain-openai==1.1.6
langchain-text-splitters==1.1.0
langchain-chroma==1.1.0
langchain-huggingface==1.2.0

# ================================
# 向量数据库和 Embedding
# ================================
chromadb==1.4.0
# 使用已知稳定的版本组合（经过兼容性测试）
sentence-transformers==3.3.1
transformers==4.45.2
tokenizers==0.20.0
torch==2.9.1

# ================================
# 前端和文档处理
# ================================
streamlit==1.52.2
pypdf==6.5.0

# ================================
# 工具库
# ================================
python-dotenv==1.2.1    # 环境变量管理
openai==2.14.0          # OpenAI API 客户端
